% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition_models.R
\name{apply_countmodel_to_column}
\alias{apply_countmodel_to_column}
\title{Apply a GLMM Count Model and Calculate Variance Partition}
\usage{
apply_countmodel_to_column(column_data, meta_data)
}
\arguments{
\item{column_data}{A numeric vector of counts (response variable).}

\item{meta_data}{A data frame containing the covariates for the model.}
}
\value{
A `varPart` object from the variancePartition package.
}
\description{
A wrapper to apply a specific lme4 generalized linear mixed model (Poisson)
and calculate the variance explained by each random effect.
}
