% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_partition_models.R
\name{apply_model_to_column}
\alias{apply_model_to_column}
\title{Apply a Mixed Model and Calculate Variance Partition}
\usage{
apply_model_to_column(column_data)
}
\arguments{
\item{column_data}{A numeric vector to be used as the response variable.}
}
\value{
A `varPart` object from the variancePartition package.
}
\description{
A wrapper to apply a specific lme4 mixed model to a column of data
and calculate the variance explained by each random effect using variancePartition.
}
