% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lisi.R
\name{compute_lisi_parallel}
\alias{compute_lisi_parallel}
\title{Compute LISI in Parallel}
\usage{
compute_lisi_parallel(
  X,
  meta_data,
  label_colnames,
  perplexity = 30,
  nn_eps = 0,
  n_thread = 1
)
}
\arguments{
\item{X}{A numeric matrix of embeddings (e.g., PCs) for which to compute LISI.}

\item{meta_data}{A data frame with cell metadata, where rows correspond to rows of X.}

\item{label_colnames}{A character vector of column names in `meta_data` for which to compute LISI.}

\item{perplexity}{The perplexity parameter used to determine the neighborhood size.}

\item{nn_eps}{A parameter for the nearest neighbor search.}

\item{n_thread}{The number of parallel threads to use.}
}
\value{
A data frame with LISI scores for each cell (rows) and each label (columns).
}
\description{
Calculates the Local Inverse Simpson's Index (LISI) for multiple metadata
labels in parallel.
}
