% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NAM.R
\name{conditional_permutation}
\alias{conditional_permutation}
\title{Conditional Permutation}
\usage{
conditional_permutation(B, Y, num)
}
\arguments{
\item{B}{A vector defining the groups.}

\item{Y}{The numeric vector to be permuted.}

\item{num}{The number of permutations to generate.}
}
\value{
A matrix with `num` columns, each representing a conditional permutation of `Y`.
}
\description{
Generates permutations of a vector, conditioned on a grouping factor.
The permutation is performed within each group defined by `B`.
}
