% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_normalize}
\alias{cosine_normalize}
\title{Cosine Normalize a Matrix}
\usage{
cosine_normalize(x, dim)
}
\arguments{
\item{x}{A numeric matrix.}

\item{dim}{The margin to apply normalization over (1 for rows, 2 for columns).}
}
\value{
A cosine-normalized matrix.
}
\description{
Normalizes vectors in a matrix to unit length (cosine normalization).
}
