% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NAM.R
\name{empirical_fdrs}
\alias{empirical_fdrs}
\title{Calculate Empirical False Discovery Rates}
\usage{
empirical_fdrs(z, znull, thresholds)
}
\arguments{
\item{z}{A vector of observed statistics.}

\item{znull}{A matrix of null statistics from permutations.}

\item{thresholds}{A numeric vector of thresholds for calculating FDR.}
}
\value{
A vector of empirical FDRs.
}
\description{
Computes empirical FDRs by comparing observed statistics to a set of null statistics
at given thresholds.
}
