% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_interaction.R
\name{generate_data_interaction}
\alias{generate_data_interaction}
\title{Generate Simulated Single-Cell Data with eQTLs}
\usage{
generate_data_interaction(
  n_cells = 3000,
  sd_celltypes = 0.1,
  n_major_cell_types = 7,
  n_minor_cell_types = 3,
  relative_abundance = 0.1,
  n_major_interact_celltypes = 1,
  n_minor_interact_celltypes = 1,
  n_individuals = 30,
  n_batchs = 4,
  prop_sex = 0.5,
  prop_disease = 0.5,
  interaction_feature_1 = c("sex", "age", "disease", "bmi"),
  interaction_feature_2 = c("sex", "age", "disease", "bmi"),
  fc_interact = 0.1,
  interaction_type = NULL,
  n_features = 20,
  cluster_features = 1:20,
  disease_features = 0,
  sex_features = 0,
  age_features = 0,
  bmi_features = 0,
  batch_features = 0,
  individual_features = 0,
  cluster_ratio = 0.25,
  disease_ratio = 0,
  sex_ratio = 0,
  age_ratio = 0,
  bmi_ratio = 0,
  batch_ratio = 0,
  individual_ratio = 0.1,
  ratio_variance = 0.5,
  cluster_col = "cell_type",
  disease_col = "disease",
  sex_col = "sex",
  age_col = "age",
  bmi_col = "bmi",
  batch_col = "batch",
  individual_col = "subject_id",
  dist_type = "nb",
  size_nb = 1,
  sparsity = 0,
  transition_prob = 0.5,
  low_gene_prop = 0.2,
  generate_genotype = TRUE,
  eqtl_celltype = "A",
  eqtl_gene = "Gene1",
  cor_vals = c(0, 0.15, 0.3),
  allele_freqs = c(0.05, 0.3),
  n_snps = 100,
  seed = 1234,
  n_thread = 1,
  verbose = TRUE
)
}
\arguments{
\item{n_cells}{Base number of cells for major cell types per individual.}

\item{sd_celltypes}{Standard deviation for the number of cells.}

\item{n_major_cell_types}{Number of major cell types to simulate.}

\item{n_minor_cell_types}{Number of minor (rare) cell types to simulate.}

\item{relative_abundance}{The abundance ratio of minor to major cell types.}

\item{n_individuals}{Total number of individuals to simulate.}

\item{n_batchs}{Number of batches to distribute individuals into.}

\item{prop_sex}{Proportion of individuals assigned to sex '1'.}

\item{prop_disease}{Proportion of individuals assigned to disease group '1'.}

\item{n_features}{The number of latent features (pseudo principal components) to generate.}

\item{cluster_features}{Integer vector specifying which features should be associated with cell type clusters.}

\item{disease_features}{Integer vector specifying which features should be associated with disease status.}

\item{sex_features}{Integer vector specifying which features should be associated with sex.}

\item{age_features}{Integer vector specifying which features should be associated with age.}

\item{bmi_features}{Integer vector specifying which features should be associated with BMI.}

\item{batch_features}{Integer vector specifying which features should be associated with batch.}

\item{individual_features}{Integer vector specifying which features should be associated with individual identity.}

\item{cluster_ratio}{Maximum proportion of variance a `cluster_feature` can explain.}

\item{disease_ratio}{Maximum proportion of variance a `disease_feature` can explain.}

\item{sex_ratio}{Maximum proportion of variance a `sex_feature` can explain.}

\item{age_ratio}{Maximum proportion of variance an `age_feature` can explain.}

\item{bmi_ratio}{Maximum proportion of variance a `bmi_feature` can explain.}

\item{batch_ratio}{Maximum proportion of variance a `batch_feature` can explain.}

\item{individual_ratio}{Maximum proportion of variance an `individual_feature` can explain.}

\item{ratio_variance}{Numeric, amount of random variation to add to the variance ratios.}

\item{cluster_col}{Name of the cell type column in the metadata.}

\item{disease_col}{Name of the disease status column in the metadata.}

\item{sex_col}{Name of the sex column in the metadata.}

\item{age_col}{Name of the age column in the metadata.}

\item{bmi_col}{Name of the BMI column in the metadata.}

\item{batch_col}{Name of the batch column in the metadata.}

\item{individual_col}{Name of the individual ID column in the metadata.}

\item{dist_type}{The distribution to use for generating counts from features ("nb" for negative binomial or "poisson").}

\item{size_nb}{Size parameter for the negative binomial distribution (only used if `dist_type` is "nb").}

\item{sparsity}{The proportion of zero counts to introduce into the simulated count data.}

\item{transition_prob}{The percentile (e.g., 0.9 for 90th percentile) to use for determining the transition to overdispersion.}

\item{low_gene_prop}{The proportion of genes to be assigned low expression levels. (only used if `dist_type` is "poisson").}

\item{generate_genotype}{Logical. If TRUE, genotypes will be simulated.}

\item{eqtl_celltype}{The cell type in which to simulate the eQTL effect.}

\item{eqtl_gene}{The gene for which to simulate the eQTL effect.}

\item{cor_vals}{A numeric vector of target correlations for the eQTL simulation.}

\item{allele_freqs}{A numeric vector of allele frequencies for the eQTL simulation.}

\item{n_snps}{The number of simulated SNPs to generate for each condition.}

\item{seed}{An integer for reproducibility.}

\item{n_thread}{The number of parallel threads to use.}

\item{verbose}{Logical. If TRUE, progress messages will be printed.}

\item{n_major_diff_celltypes}{Number of major cell types to set as differentially abundant in the 'disease' group.}

\item{n_minor_diff_celltypes}{Number of minor cell types to set as differentially abundant in the 'disease' group.}

\item{fc_increase}{The fold change increase for differentially abundant cell types in the disease group.}
}
\value{
A list containing several data frames: `simulation_counts` (genes x cells count matrix),
  `dummy_data` (cell-level metadata), `bulk_df` (pseudobulk data),
  `genotype_df_list` (a list of simulated genotype matrices), and
  `cor_summary_df_list` (a summary of the simulated eQTL correlations).
}
\description{
This is the main wrapper function to generate a complex simulated dataset.
It simulates cell metadata, gene expression counts based on latent features,
and pseudo-genotypes correlated with the expression of a specific gene in a
specific cell type to simulate eQTL effects.
}
