% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data.R
\name{generate_dummy_data}
\alias{generate_dummy_data}
\title{Generate Dummy Cell Metadata}
\usage{
generate_dummy_data(
  n_cells = 3000,
  sd_celltypes = 0.1,
  n_major_cell_types = 7,
  n_minor_cell_types = 3,
  relative_abundance = 0.1,
  n_major_diff_celltypes = 1,
  n_minor_diff_celltypes = 1,
  n_individuals = 30,
  n_batchs = 4,
  prop_sex = 0.5,
  prop_disease = 0.5,
  fc_increase = 0.1,
  seed = 1234
)
}
\arguments{
\item{n_cells}{Base number of cells for major cell types per individual.}

\item{sd_celltypes}{Standard deviation for the number of cells.}

\item{n_major_cell_types}{Number of major cell types to simulate.}

\item{n_minor_cell_types}{Number of minor (rare) cell types to simulate.}

\item{relative_abundance}{The abundance ratio of minor to major cell types.}

\item{n_major_diff_celltypes}{Number of major cell types to make differentially abundant.}

\item{n_minor_diff_celltypes}{Number of minor cell types to make differentially abundant.}

\item{n_individuals}{Total number of individuals to simulate.}

\item{n_batchs}{Number of batches to distribute individuals into.}

\item{prop_sex}{Proportion of individuals assigned to sex '1'.}

\item{prop_disease}{Proportion of individuals assigned to disease group '1'.}

\item{fc_increase}{The fold change increase for differentially abundant cell types.}

\item{seed}{An integer for reproducibility.}
}
\value{
A list containing two elements: a data frame of the simulated cell metadata,
  and a character vector of the cell types designated as differentially abundant.
}
\description{
Simulates cell-level metadata including individual-level covariates
(sex, disease, age, batch, bmi) and cell type assignments. It can introduce
differential cell type abundance based on disease status.
}
