% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_genotype.R
\name{generate_genotype}
\alias{generate_genotype}
\title{Generate Simulated Single-Cell Data with eQTLs}
\usage{
generate_genotype(
  bulk_df,
  meta_data,
  eqtl_celltype = "A",
  eqtl_gene = "Gene1",
  cor_vals = c(0, 0.15, 0.3),
  allele_freqs = c(0.05, 0.3),
  n_snps = 100,
  seed = 1234,
  n_thread = 1,
  verbose = TRUE
)
}
\arguments{
\item{bulk_df}{A data frame containing pseudobulk expression data with columns for cell type and gene expression.}

\item{eqtl_celltype}{The cell type in which to simulate the eQTL effect.}

\item{eqtl_gene}{The gene for which to simulate the eQTL effect.}

\item{cor_vals}{A numeric vector of target correlations for the eQTL simulation.}

\item{allele_freqs}{A numeric vector of allele frequencies for the eQTL simulation.}

\item{n_snps}{The number of simulated SNPs to generate for each condition.}

\item{seed}{An integer for reproducibility.}

\item{n_thread}{The number of parallel threads to use.}

\item{verbose}{Logical. If TRUE, progress messages will be printed.}
}
\value{
A list containing several data frames: `genotype_df_list` (a list of simulated genotype matrices), and
  `cor_summary_df_list` (a summary of the simulated eQTL correlations).
}
\description{
This is a function to generate genotype data using provided pseudobulk expression data.
}
