% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NormalizeDataSeurat}
\alias{NormalizeDataSeurat}
\title{Normalize data using Seurat's log-transform method}
\usage{
NormalizeDataSeurat(A, scaling_factor = 10000, do_ftt = FALSE)
}
\arguments{
\item{A}{A sparse matrix of counts.}

\item{scaling_factor}{A scaling factor, defaults to 10000.}

\item{do_ftt}{Logical, if TRUE, applies a Freeman-Tukey stabilizing transformation.}
}
\value{
A log-normalized sparse matrix.
}
\description{
A standalone implementation of Seurat's LogNormalize method.
}
