% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NAM.R
\name{tail_counts}
\alias{tail_counts}
\title{Tail Counts for FDR Calculation}
\usage{
tail_counts(z, znull)
}
\arguments{
\item{z}{A vector of thresholds (derived from observed statistics).}

\item{znull}{A matrix of null statistics.}
}
\value{
A matrix where each column corresponds to a column of `znull` and rows
represent the count of null values exceeding the squared thresholds in `z`.
}
\description{
A helper function to count the number of null statistics exceeding a set of thresholds.
This is used in the empirical FDR calculation.
}
