% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unregister_dopar}
\alias{unregister_dopar}
\title{Unregister a doParallel Backend}
\usage{
unregister_dopar()
}
\value{
Invisible
}
\description{
A helper function to clean up and unregister a parallel backend created by
the `foreach` package.
}
